Changed some Integers to Floats.

+++ ppm/naive_ppm.py
@@ -30,65 +30,65 @@ def Portfolio_Performance_Summary(sdf, VAR, PER):
         .withColumn(f"Balance_{PER}", (f.col("Balance") / f.col("Records")).astype(t.IntegerType())) \
         .withColumn(f"Instalment_{PER}", (f.col("Instalment") / f.col("Records")).astype(t.IntegerType())) \
         .withColumn(f"NextCycle30Days_{PER}",
-                    (f.col("NextCycle30Days") / f.col("Records") * f.lit(100)).astype(t.IntegerType())) \
+                    (f.col("NextCycle30Days") / f.col("Records") * f.lit(100)).astype(t.FloatType())) \
         .withColumn(f"SecondCycle60Days_{PER}",
-                    (f.col("SecondCycle60Days") / f.col("Records") * f.lit(100)).astype(t.IntegerType())) \
+                    (f.col("SecondCycle60Days") / f.col("Records") * f.lit(100)).astype(t.FloatType())) \
         .withColumn(f"ThirdCycle90Days_{PER}",
-                    (f.col("ThirdCycle90Days") / f.col("Records") * f.lit(100)).astype(t.IntegerType())) \
+                    (f.col("ThirdCycle90Days") / f.col("Records") * f.lit(100)).astype(t.FloatType())) \
         .withColumn(f"BadRate1M_{PER}",
                     ((f.col("Bads1") / reduce(add, [f.col("Goods1"), f.col("Bads1")])) * f.lit(100)).astype(
-                        t.IntegerType())) \
+                        t.FloatType())) \
         .withColumn(f"BadRate3M_{PER}",
                     ((f.col("Bads3") / reduce(add, [f.col("Goods3"), f.col("Bads3")])) * f.lit(100)).astype(
-                        t.IntegerType())) \
+                        t.FloatType())) \
         .withColumn(f"BadRate6M_{PER}",
                     ((f.col("Bads6") / reduce(add, [f.col("Goods6"), f.col("Bads6")])) * f.lit(100)).astype(
-                        t.IntegerType())) \
+                        t.FloatType())) \
         .withColumn(f"BadRate9M_{PER}",
                     ((f.col("Bads9") / reduce(add, [f.col("Goods9"), f.col("Bads9")])) * f.lit(100)).astype(
-                        t.IntegerType())) \
+                        t.FloatType())) \
         .withColumn(f"BadRate12M_{PER}",
                     ((f.col("Bads12") / reduce(add, [f.col("Goods12"), f.col("Bads12")])) * f.lit(100)).astype(
-                        t.IntegerType())) \
+                        t.FloatType())) \
         .withColumn(f"DoubtfulDebt1M_{PER}",
-                    (f.col("DoubtfulDebt1M") / f.col("Records") * f.lit(100)).astype(t.IntegerType())) \
+                    (f.col("DoubtfulDebt1M") / f.col("Records") * f.lit(100)).astype(t.FloatType())) \
         .withColumn(f"DoubtfulDebt3M_{PER}",
-                    (f.col("DoubtfulDebt3M") / f.col("Records") * f.lit(100)).astype(t.IntegerType())) \
+                    (f.col("DoubtfulDebt3M") / f.col("Records") * f.lit(100)).astype(t.FloatType())) \
         .withColumn(f"DoubtfulDebt6M_{PER}",
-                    (f.col("DoubtfulDebt6M") / f.col("Records") * f.lit(100)).astype(t.IntegerType())) \
+                    (f.col("DoubtfulDebt6M") / f.col("Records") * f.lit(100)).astype(t.FloatType())) \
         .withColumn(f"DoubtfulDebt9M_{PER}",
-                    (f.col("DoubtfulDebt9M") / f.col("Records") * f.lit(100)).astype(t.IntegerType())) \
+                    (f.col("DoubtfulDebt9M") / f.col("Records") * f.lit(100)).astype(t.FloatType())) \
         .withColumn(f"DoubtfulDebt12M_{PER}",
-                    (f.col("DoubtfulDebt12M") / f.col("Records") * f.lit(100)).astype(t.IntegerType())) \
+                    (f.col("DoubtfulDebt12M") / f.col("Records") * f.lit(100)).astype(t.FloatType())) \
         .withColumn(f"Records_DC0_{PER}", f.col("Records_DC0").astype(t.IntegerType())) \
         .withColumn(f"Records_DC1_{PER}", f.col("Records_DC1").astype(t.IntegerType())) \
         .withColumn(f"Records_DC2_{PER}", f.col("Records_DC2").astype(t.IntegerType())) \
         .withColumn(f"Records_DC3_{PER}", f.col("Records_DC3").astype(t.IntegerType())) \
         .withColumn(f"Records_DC4_{PER}", f.col("Records_DC4").astype(t.IntegerType())) \
         .withColumn(f"DC0_Credit_Paidup_{PER}",
-                    (f.col("DC0_Credit_Paidup") / f.col("Records_DC0") * f.lit(100)).astype(t.IntegerType())) \
-        .withColumn(f"DC0_UTD_{PER}", (f.col("DC0_UTD") / f.col("Records_DC0") * f.lit(100)).astype(t.IntegerType())) \
+                    (f.col("DC0_Credit_Paidup") / f.col("Records_DC0") * f.lit(100)).astype(t.FloatType())) \
+        .withColumn(f"DC0_UTD_{PER}", (f.col("DC0_UTD") / f.col("Records_DC0") * f.lit(100)).astype(t.FloatType())) \
         .withColumn(f"DC0_Forward_Roll_{PER}",
-                    (f.col("DC0_Forward_Roll") / f.col("Records_DC0") * f.lit(100)).astype(t.IntegerType())) \
-        .withColumn(f"DC1_Cured_{PER}", (f.col("DC1_Cured") / f.col("Records_DC1") * f.lit(100)).astype(t.IntegerType())) \
-        .withColumn(f"DC1_30Days_{PER}", (f.col("DC1_30Days") / f.col("Records_DC1") * f.lit(100)).astype(t.IntegerType())) \
+                    (f.col("DC0_Forward_Roll") / f.col("Records_DC0") * f.lit(100)).astype(t.FloatType())) \
+        .withColumn(f"DC1_Cured_{PER}", (f.col("DC1_Cured") / f.col("Records_DC1") * f.lit(100)).astype(t.FloatType())) \
+        .withColumn(f"DC1_30Days_{PER}", (f.col("DC1_30Days") / f.col("Records_DC1") * f.lit(100)).astype(t.FloatType())) \
         .withColumn(f"DC1_Forward_Roll_{PER}",
-                    (f.col("DC1_Forward_Roll") / f.col("Records_DC1") * f.lit(100)).astype(t.IntegerType())) \
-        .withColumn(f"DC2_Cured_{PER}", (f.col("DC2_Cured") / f.col("Records_DC2") * f.lit(100)).astype(t.IntegerType())) \
+                    (f.col("DC1_Forward_Roll") / f.col("Records_DC1") * f.lit(100)).astype(t.FloatType())) \
+        .withColumn(f"DC2_Cured_{PER}", (f.col("DC2_Cured") / f.col("Records_DC2") * f.lit(100)).astype(t.FloatType())) \
         .withColumn(f"DC2_Backward_Roll_{PER}",
-                    (f.col("DC2_Backward_Roll") / f.col("Records_DC2") * f.lit(100)).astype(t.IntegerType())) \
-        .withColumn(f"DC2_60Days_{PER}", (f.col("DC2_60Days") / f.col("Records_DC2") * f.lit(100)).astype(t.IntegerType())) \
+                    (f.col("DC2_Backward_Roll") / f.col("Records_DC2") * f.lit(100)).astype(t.FloatType())) \
+        .withColumn(f"DC2_60Days_{PER}", (f.col("DC2_60Days") / f.col("Records_DC2") * f.lit(100)).astype(t.FloatType())) \
         .withColumn(f"DC2_Forward_Roll_{PER}",
-                    (f.col("DC2_Forward_Roll") / f.col("Records_DC2") * f.lit(100)).astype(t.IntegerType())) \
-        .withColumn(f"DC3_Cured_{PER}", (f.col("DC3_Cured") / f.col("Records_DC3") * f.lit(100)).astype(t.IntegerType())) \
+                    (f.col("DC2_Forward_Roll") / f.col("Records_DC2") * f.lit(100)).astype(t.FloatType())) \
+        .withColumn(f"DC3_Cured_{PER}", (f.col("DC3_Cured") / f.col("Records_DC3") * f.lit(100)).astype(t.FloatType())) \
         .withColumn(f"DC3_Backward_Roll_{PER}",
-                    (f.col("DC3_Backward_Roll") / f.col("Records_DC3") * f.lit(100)).astype(t.IntegerType())) \
-        .withColumn(f"DC3_90Days_{PER}", (f.col("DC3_90Days") / f.col("Records_DC3") * f.lit(100)).astype(t.IntegerType())) \
+                    (f.col("DC3_Backward_Roll") / f.col("Records_DC3") * f.lit(100)).astype(t.FloatType())) \
+        .withColumn(f"DC3_90Days_{PER}", (f.col("DC3_90Days") / f.col("Records_DC3") * f.lit(100)).astype(t.FloatType())) \
         .withColumn(f"DC3_Forward_Roll_{PER}",
-                    (f.col("DC3_Forward_Roll") / f.col("Records_DC3") * f.lit(100)).astype(t.IntegerType())) \
+                    (f.col("DC3_Forward_Roll") / f.col("Records_DC3") * f.lit(100)).astype(t.FloatType())) \
         .withColumn(f"DC4_Backward_Roll_{PER}",
-                    (f.col("DC4_Backward_Roll") / f.col("Records_DC4") * f.lit(100)).astype(t.IntegerType())) \
-        .withColumn(f"DC4_120Days_{PER}", (f.col("DC4_120Days") / f.col("Records_DC4") * f.lit(100)).astype(t.IntegerType()))
+                    (f.col("DC4_Backward_Roll") / f.col("Records_DC4") * f.lit(100)).astype(t.FloatType())) \
+        .withColumn(f"DC4_120Days_{PER}", (f.col("DC4_120Days") / f.col("Records_DC4") * f.lit(100)).astype(t.FloatType()))
 
     sdf_3 = sdf_2\
         .drop(*ls_summary_vars)
